requirements.txt
fastapi>=0.100
uvicorn[standard]>=0.20
pydantic>=2.0
requests>=2.28
openai>=1.0
python-dotenv>=1.0
